Move static repo methods to extension methods for tests
